inital rename of BodyCopy to BodyText. Lots of broken stuff.
