+++ src/@ui/typography/BodyText/index.js
@@ -36,7 +36,7 @@ const styles = styled(({ theme, bold, italic }) => {
   });
 }, 'BodyCopy');
 
-const BodyCopy = compose(
+const BodyText = compose(
   setPropTypes({
     bold: PropTypes.bool,
     italic: PropTypes.bool,
@@ -53,4 +53,4 @@ const BodyCopy = compose(
   pure,
 )(Text);
 
-export default BodyCopy;
+export default BodyText;
