+++ src/@ui/typography/index.js
@@ -5,5 +5,5 @@ export { default as H4 } from './H4';
 export { default as H5 } from './H5';
 export { default as H6 } from './H6';
 export { default as H7 } from './H7';
-export { default as BodyCopy } from './BodyCopy';
+export { default as BodyText } from './BodyText';
 export { default as UIText } from './UIText';
